Rename project Steeltoe.Logging.DynamicLogger to Steeltoe.Logging.DynamicConsole
