+++ src/Bootstrap/src/AutoConfiguration/PublicAPI.Unshipped.txt
@@ -9,7 +9,7 @@ const Steeltoe.Bootstrap.AutoConfiguration.SteeltoeAssemblyNames.Connectors = "S
 const Steeltoe.Bootstrap.AutoConfiguration.SteeltoeAssemblyNames.DiscoveryConfiguration = "Steeltoe.Discovery.Configuration" -> string!
 const Steeltoe.Bootstrap.AutoConfiguration.SteeltoeAssemblyNames.DiscoveryConsul = "Steeltoe.Discovery.Consul" -> string!
 const Steeltoe.Bootstrap.AutoConfiguration.SteeltoeAssemblyNames.DiscoveryEureka = "Steeltoe.Discovery.Eureka" -> string!
-const Steeltoe.Bootstrap.AutoConfiguration.SteeltoeAssemblyNames.LoggingDynamicLogger = "Steeltoe.Logging.DynamicLogger" -> string!
+const Steeltoe.Bootstrap.AutoConfiguration.SteeltoeAssemblyNames.LoggingDynamicConsole = "Steeltoe.Logging.DynamicConsole" -> string!
 const Steeltoe.Bootstrap.AutoConfiguration.SteeltoeAssemblyNames.LoggingDynamicSerilog = "Steeltoe.Logging.DynamicSerilog" -> string!
 const Steeltoe.Bootstrap.AutoConfiguration.SteeltoeAssemblyNames.ManagementEndpoint = "Steeltoe.Management.Endpoint" -> string!
 const Steeltoe.Bootstrap.AutoConfiguration.SteeltoeAssemblyNames.ManagementPrometheus = "Steeltoe.Management.Prometheus" -> string!
