+++ src/Logging/src/DynamicConsole/PublicAPI.Unshipped.txt
@@ -1,0 +1,5 @@
+#nullable enable
+static Steeltoe.Logging.DynamicConsole.LoggingBuilderExtensions.AddDynamicConsole(this Microsoft.Extensions.Logging.ILoggingBuilder! builder) -> Microsoft.Extensions.Logging.ILoggingBuilder!
+Steeltoe.Logging.DynamicConsole.DynamicConsoleLoggerProvider
+Steeltoe.Logging.DynamicConsole.DynamicConsoleLoggerProvider.DynamicConsoleLoggerProvider(Microsoft.Extensions.Options.IOptionsMonitor<Microsoft.Extensions.Logging.LoggerFilterOptions!>! filterOptionsMonitor, Microsoft.Extensions.Logging.Console.ConsoleLoggerProvider! consoleLoggerProvider, System.Collections.Generic.IEnumerable<Steeltoe.Logging.IDynamicMessageProcessor!>! messageProcessors) -> void
+Steeltoe.Logging.DynamicConsole.LoggingBuilderExtensions
