+++ src/Logging/src/DynamicConsole/Steeltoe.Logging.DynamicConsole.csproj
