+++ src/Steeltoe.Configuration.slnf
@@ -29,7 +29,7 @@
       "Discovery\\src\\Consul\\Steeltoe.Discovery.Consul.csproj",
       "Discovery\\src\\Eureka\\Steeltoe.Discovery.Eureka.csproj",
       "Logging\\src\\Abstractions\\Steeltoe.Logging.Abstractions.csproj",
-      "Logging\\src\\DynamicLogger\\Steeltoe.Logging.DynamicLogger.csproj",
+      "Logging\\src\\DynamicConsole\\Steeltoe.Logging.DynamicConsole.csproj",
       "Management\\src\\Abstractions\\Steeltoe.Management.Abstractions.csproj",
       "Management\\src\\Endpoint\\Steeltoe.Management.Endpoint.csproj",
       "Tools\\src\\ConfigurationSchemaGenerator\\ConfigurationSchemaGenerator.csproj"
